transferindo arquivos de 3G

+++ src/include/server_utils.h
@@ -20,5 +20,5 @@
 char *build_file_path(char* file_name, char *dir_path);
 int read_from_client(int sock, char *buf, int len);
 int write_to_client(int sock, char *buf, int length);
-void print_header(int sock, int number_of_threads, int file_size);
+void print_header(int sock, int number_of_threads, long file_size);
 #endif /* SERVER_UTILS_H_ */
